requirements.txt
streamlit==1.26.0
pandas==2.2.2
numpy==2.2.2
scikit-learn==1.2.2
joblib==1.3.2
pillow==10.1.0
xgboost==1.7.6